@@ -30,20 +34,22 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "react-ga": "2.5.3",
-    "remove-markdown": "0.2.2",
-    "shelljs": "0.7.8"
+    "intl": "^1.2.5",
+    "react-ga": "^2.4.1",
+    "remove-markdown": "^0.2.2",
+    "shelljs": "^0.7.8"
   },
   "deprecated": false,
   "description": "Strapi Admin",
   "devDependencies": {
-    "sanitize.css": "4.1.0",
-    "strapi-helper-plugin": "3.0.0-alpha.14.3",
-    "strapi-utils": "3.0.0-alpha.14.3"
+    "cross-env": "^5.0.5",
+    "sanitize.css": "^4.1.0",
+    "strapi-helper-plugin": "3.0.0-alpha.24.1",
+    "strapi-utils": "3.0.0-alpha.24.1"
   },
   "engines": {
-    "node": ">= 9.0.0",
-    "npm": ">= 5.0.0"
+    "node": ">= 10.0.0",
+    "npm": ">= 6.0.0"
   },
   "homepage": "https://github.com/strapi/strapi-admin#readme",
   "license": "MIT",
