@@ -47,8 +47,22 @@ class LeftMenuLink extends React.Component {
       <span className={styles.linkLabel}>{this.props.label}</span>
     );
 
-    return (
-      <li className={styles.item}>
+    // Icon.
+    const icon = <i className={`${styles.linkIcon} fa-${this.props.icon} fa`} />;
+
+    // Create external or internal link.
+    const link = this.props.destination.includes('http')
+      ? (
+        <a
+          className={`${styles.link} ${isLinkActive ? styles.linkActive : ''}`}
+          href={this.props.destination}
+          target="_blank"
+        >
+          {icon}
+          {content}
+        </a>
+      )
+      : (
         <Link
           className={`${styles.link} ${isLinkActive ? styles.linkActive : ''}`}
           to={{
@@ -56,9 +70,14 @@ class LeftMenuLink extends React.Component {
             search: this.props.source ? `?source=${this.props.source}` : '',
           }}
         >
-          <i className={`${styles.linkIcon} fa-${this.props.icon} fa`} />
+          {icon}
           {content}
         </Link>
+      );
+
+    return (
+      <li className={styles.item}>
+        {link}
         {plugin}
       </li>
     );
