@@ -10,30 +10,23 @@ import cn from 'classnames';
 import { isEmpty, replace } from 'lodash';
 import { FormattedMessage } from 'react-intl';
 
-// Temporary picture
 import Button from 'components/Button';
 import InstallPluginPopup from 'components/InstallPluginPopup';
-import Official from 'components/Official';
-// import StarsContainer from 'components/StarsContainer';
 
-import logoTShirt from 'assets/images/logo-t-shirt.svg';
 import styles from './styles.scss';
-import Screenshot from './screenshot.png';
+
+const PLUGINS_WITH_CONFIG = ['content-manager', 'email', 'upload'];
 
 /* eslint-disable react/no-unused-state */
 class PluginCard extends React.Component {
-  state = { isOpen: false, boostrapCol: 'col-lg-4' };
+  state = { 
+    boostrapCol: 'col-lg-4',
+  };
 
   componentDidMount() {
-    this.shouldOpenModal(this.props);
+    // Listen window resize.
     window.addEventListener('resize', this.setBoostrapCol);
     this.setBoostrapCol();
-  }
-
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.history.location.hash !== this.props.history.location.hash) {
-      this.shouldOpenModal(nextProps);
-    }
   }
 
   componentWillUnmount() {
@@ -65,6 +58,15 @@ class PluginCard extends React.Component {
     }
   }
 
+  handleClickSettings = (e) => {
+    const settingsPath = this.props.plugin.id === 'content-manager' ? '/plugins/content-manager/ctm-configurations' : `/plugins/${this.props.plugin.id}/configurations/${this.props.currentEnvironment}`; 
+
+    e.preventDefault();
+    e.stopPropagation();
+
+    this.props.history.push(settingsPath);
+  }
+
   handleDownloadPlugin = (e) => {
     if (!this.props.isAlreadyInstalled && this.props.plugin.id !== 'support-us') {
       this.props.downloadPlugin(e);
@@ -75,23 +77,15 @@ class PluginCard extends React.Component {
     }
   }
 
-  shouldOpenModal = (props) => {
-    this.setState({ isOpen: !isEmpty(props.history.location.hash) });
-  }
-
   render() {
-    const buttonClass = !this.props.isAlreadyInstalled || this.props.showSupportUsButton ? styles.primary : styles.secondary;
+    const buttonClass = !this.props.isAlreadyInstalled ? styles.primary : styles.secondary;
+    const buttonLabel = this.props.isAlreadyInstalled ? 'app.components.PluginCard.Button.label.install' : 'app.components.PluginCard.Button.label.download';
 
-    let buttonLabel = this.props.isAlreadyInstalled ? 'app.components.PluginCard.Button.label.install' : 'app.components.PluginCard.Button.label.download';
-
-    if (this.props.showSupportUsButton) {
-      buttonLabel = 'app.components.PluginCard.Button.label.support';
-    }
-
-    const pluginIcon = (
-      <div className={styles.frame}>
-        <span className={styles.helper} />
-        <img src={`${this.props.plugin.id === 'support-us' ? logoTShirt : this.props.plugin.logo}`} alt="icon" />
+    // Display settings link for a selection of plugins.
+    const settingsComponent = (PLUGINS_WITH_CONFIG.includes(this.props.plugin.id) && 
+      <div className={styles.settings} onClick={this.handleClickSettings}>
+        <i className='fa fa-cog' />
+        <FormattedMessage id='app.components.PluginCard.settings' />
       </div>
     );
 
@@ -101,37 +95,21 @@ class PluginCard extends React.Component {
     };
 
     return (
-      <div className={cn(this.state.boostrapCol, styles.pluginCard)} onClick={this.handleClick}>
+      <div className={cn(this.state.boostrapCol, styles.pluginCard)}>
         <div className={styles.wrapper}>
           <div className={styles.cardTitle}>
-            {pluginIcon}
-            <div>{this.props.plugin.name}</div>
+            <div className={styles.frame}>
+              <span className={styles.helper} />
+              <img src={this.props.plugin.logo} alt="icon" />
+            </div>
+            <div>{this.props.plugin.name} <i className='fa fa-external-link' onClick={() => window.open(`https://github.com/strapi/strapi/tree/master/packages/strapi-plugin-${this.props.plugin.id}`, '_blank')} /></div>
           </div>
           <div className={styles.cardDescription}>
-            {descriptions.short}
-            &nbsp;<FormattedMessage id="app.components.PluginCard.more-details" />
-          </div>
-          <div className={styles.cardScreenshot} style={{ backgroundImage: `url(${Screenshot})` }}>
-
-          </div>
-          <div className={styles.cardPrice}>
-            <div>
-              <i className={`fa fa-${this.props.plugin.isCompatible ? 'check' : 'times'}`} />
-              <FormattedMessage id={`app.components.PluginCard.compatible${this.props.plugin.id === 'support-us' ? 'Community' : ''}`} />
-            </div>
-            <div>{this.props.plugin.price !== 0 ? `${this.props.plugin.price}€` : ''}</div>
+            {descriptions.long}
+            {/* &nbsp;<FormattedMessage id="app.components.PluginCard.more-details" /> */}
           </div>
           <div className={styles.cardFooter} onClick={e => e.stopPropagation()}>
-            <div className={styles.ratings}>
-              {/*<StarsContainer ratings={this.props.plugin.ratings} />
-              <div>
-                <span style={{ fontWeight: '600', color: '#333740' }}>{this.props.plugin.ratings}</span>
-                <span style={{ fontWeight: '500', color: '#666666' }}>/5</span>
-              </div>
-              */}
-              <Official />
-            </div>
-            <div>
+            <div className={styles.cardFooterButton}>
               <Button
                 className={cn(buttonClass, styles.button)}
                 label={buttonLabel}
@@ -146,6 +124,18 @@ class PluginCard extends React.Component {
                 &nbsp;
               </a>
             </div>
+            {this.props.isAlreadyInstalled ? 
+              (
+                settingsComponent
+              )
+              :
+              (
+                <div className={styles.compatible}>
+                  <i className={`fa fa-${this.props.plugin.isCompatible ? 'check' : 'times'}`} />
+                  <FormattedMessage id={`app.components.PluginCard.compatible${this.props.plugin.id === 'support-us' ? 'Community' : ''}`} />
+                </div>
+              )
+            }
           </div>
         </div>
         <InstallPluginPopup
@@ -168,15 +158,14 @@ PluginCard.defaultProps = {
     price: 0,
     ratings: 5,
   },
-  showSupportUsButton: false,
 };
 
 PluginCard.propTypes = {
+  currentEnvironment: PropTypes.string.isRequired,
   downloadPlugin: PropTypes.func.isRequired,
   history: PropTypes.object.isRequired,
   isAlreadyInstalled: PropTypes.bool,
   plugin: PropTypes.object,
-  showSupportUsButton: PropTypes.bool,
 };
 
 export default PluginCard;
