@@ -4,10 +4,11 @@ import {
   ENABLE_GLOBAL_OVERLAY_BLOCKER,
 } from 'constants/overlayBlocker';
 
-import { fromJS } from 'immutable';
+import { fromJS, List } from 'immutable';
 
 import {
   FREEZE_APP,
+  GET_APP_PLUGINS_SUCCEEDED,
   PLUGIN_DELETED,
   PLUGIN_LOADED,
   UNFREEZE_APP,
@@ -16,8 +17,11 @@ import {
 } from './constants';
 
 const initialState = fromJS({
+  appPlugins: List([]),
   blockApp: false,
+  overlayBlockerData: null,
   hasUserPlugin: true,
+  isAppLoading: true,
   plugins: {},
   showGlobalAppBlocker: true,
 });
@@ -29,7 +33,19 @@ function appReducer(state = initialState, action) {
     case ENABLE_GLOBAL_OVERLAY_BLOCKER:
       return state.set('showGlobalAppBlocker', true);
     case FREEZE_APP:
-      return state.set('blockApp', true);
+      return state
+        .set('blockApp', true)
+        .update('overlayBlockerData', () => {
+          if (action.data) {
+            return action.data;
+          }
+
+          return null;
+        });
+    case GET_APP_PLUGINS_SUCCEEDED:
+      return state
+        .update('appPlugins', () => List(action.appPlugins))
+        .update('isAppLoading', () => false);
     case PLUGIN_LOADED:
       return state.setIn(['plugins', action.plugin.id], fromJS(action.plugin));
     case UPDATE_PLUGIN:
@@ -37,7 +53,9 @@ function appReducer(state = initialState, action) {
     case PLUGIN_DELETED:
       return state.deleteIn(['plugins', action.plugin]);
     case UNFREEZE_APP:
-      return state.set('blockApp', false);
+      return state
+        .set('blockApp', false)
+        .set('overlayBlockerData', null);
     case UNSET_HAS_USERS_PLUGIN:
       return state.set('hasUserPlugin', false);
     default:
