@@ -2,13 +2,13 @@
  * Combine all reducers in this file and export the combined reducers.
  */
 
-import { fromJS } from 'immutable'
-import { combineReducers } from 'redux-immutable'
-import { LOCATION_CHANGE } from 'react-router-redux'
+import { fromJS } from 'immutable';
+import { combineReducers } from 'redux-immutable';
+import { LOCATION_CHANGE } from 'react-router-redux';
 
-import globalReducer from 'containers/App/reducer'
-import languageProviderReducer from 'containers/LanguageProvider/reducer'
-import notificationProviderReducer from 'containers/NotificationProvider/reducer'
+import globalReducer from 'containers/App/reducer';
+import languageProviderReducer from 'containers/LanguageProvider/reducer';
+import notificationProviderReducer from 'containers/NotificationProvider/reducer';
 
 /*
  * routeReducer
@@ -20,33 +20,33 @@ import notificationProviderReducer from 'containers/NotificationProvider/reducer
 
 // Initial routing state
 const routeInitialState = fromJS({
-  location: null
-})
+  location: null,
+});
 
 /**
  * Merge route into the global application state
  */
-function routeReducer (state = routeInitialState, action) {
+function routeReducer(state = routeInitialState, action) {
   switch (action.type) {
     /* istanbul ignore next */
     case LOCATION_CHANGE:
       return state.merge({
-        location: action.payload
-      })
+        location: action.payload,
+      });
     default:
-      return state
+      return state;
   }
 }
 
 /**
  * Creates the main reducer with the dynamically injected ones
  */
-export default function createReducer (injectedReducers) {
+export default function createReducer(injectedReducers) {
   return combineReducers({
     route: routeReducer,
     app: globalReducer,
     language: languageProviderReducer,
     notification: notificationProviderReducer,
-    ...injectedReducers
-  })
+    ...injectedReducers,
+  });
 }
