@@ -21,23 +21,19 @@ shell.rm('-rf', path.resolve(appPath, 'package-lock.json'));
 shell.rm('-rf', path.resolve(appPath, 'admin', 'package-lock.json'));
 
 // Install the project dependencies.
-shell.exec(`cd "${appPath}" && npm install --ignore-scripts`, {
-  silent
-});
+shell.cd(appPath);
+shell.exec('npm install --ignore-scripts', {silent});
 
 // Install the administration dependencies.
-shell.exec(`cd "${path.resolve(appPath, 'admin')}" && npm install`, {
-  silent
-});
+shell.cd(path.resolve(appPath, 'admin'));
+shell.exec('npm install', {silent});
 
 if (isDevelopmentMode) {
-  shell.exec(`cd "${path.resolve(appPath, 'admin')}" && npm link strapi-helper-plugin && npm link strapi-utils`, {
-    silent
-  });
+  shell.cd(path.resolve(appPath, 'admin'));
+  shell.exec('npm link strapi-helper-plugin && npm link strapi-utils', {silent});
 } else {
-  shell.exec(`cd "${path.resolve(appPath, 'admin', 'node_modules', 'strapi-helper-plugin')}" && npm install`, {
-    silent
-  });
+  shell.cd(path.resolve(appPath, 'admin', 'node_modules', 'strapi-helper-plugin'));
+  shell.exec('npm install', {silent});
 }
 
 shell.echo('Packaged installed successfully');
